// src/typings/prometheus.d.ts
export class Registry {
  /**
   * Get string representation for all metrics
   */
  metrics(): string;

  /**
   * Remove all metrics from the registry
   */
  clear(): void;

  /**
   * Reset all metrics in the registry
   */
  resetMetrics(): void;

  /**
   * Register metric to register
   * @param metric Metric to add to register
   */
  registerMetric<T extends string>(metric: Metric<T>): void;

  /**
   * Get all metrics as objects
   */
  getMetricsAsJSON(): metric[];

  /**
   * Get all metrics as objects
   */
  getMetricsAsArray(): metric[];

  /**
   * Remove a single metric
   * @param name The name of the metric to remove
   */
  removeSingleMetric(name: string): void;

  /**
   * Get a single metric
   * @param name The name of the metric
   */
  getSingleMetric<T extends string>(name: string): Metric<T>;

  /**
   * Set static labels to every metric emitted by this registry
   * @param labels of name/value pairs:
   * { defaultLabel: "value", anotherLabel: "value 2" }
   */
  setDefaultLabels(labels: Object): void;

  /**
   * Get a string representation of a single metric by name
   * @param name The name of the metric
   */
  getSingleMetricAsString(name: string): string;

  /**
   * Gets the Content-Type of the metrics for use in the response headers.
   */
  contentType: string;

  /**
   * Merge registers
   * @param registers The registers you want to merge together
   */
  static merge(registers: Registry[]): Registry;
}
// ...
export type Metric<T extends string> = Counter<T> | Gauge<T> | Summary<T> | Histogram<T>;

/**
 * Aggregation methods, used for aggregating metrics in a Node.js cluster.
 */
export type Aggregator = 'omit' | 'sum' | 'first' | 'min' | 'max' | 'average';

export enum MetricType {
  Counter,
  Gauge,
  Histogram,
  Summary
}

interface metric {
  name: string;
  help: string;
  type: MetricType;
  aggregator: Aggregator;
}

type LabelValues<T extends string> = Partial<Record<T, string | number>>;

export interface CounterConfiguration<T extends string> {
  name: string;
  help: string;
  labelNames?: T[];
  registers?: Registry[];
  aggregator?: Aggregator;
}

/**
 * A counter is a cumulative metric that represents a single numerical value that only ever goes up
 */
export class Counter<T extends string> {
  /**
   * @param configuration Configuration when creating a Counter metric. Name and Help is required.
   */
  constructor(configuration: CounterConfiguration<T>);

  /**
   * Increment for given labels
   * @param labels Object with label keys and values
   * @param value The number to increment with
   */
  inc(labels: LabelValues<T>, value?: number): void;

  /**
   * Increment with value
   * @param value The value to increment with
   */
  inc(value?: number): void;

  /**
   * Return the child for given labels
   * @param values Label values
   * @return Configured counter with given labels
   */
  labels(...values: string[]): Counter.Internal;

  /**
   * Reset counter values
   */
  reset(): void;

  /**
   * Remove metrics for the given label values
   * @param values Label values
   */
  remove(...values: string[]): void;
}

export namespace Counter {
  interface Internal {
    /**
     * Increment with value
     * @param value The value to increment with
     */
    inc(value?: number): void;
  }
}

export interface GaugeConfiguration<T extends string> {
  name: string;
  help: string;
  labelNames?: T[];
  registers?: Registry[];
  aggregator?: Aggregator;
}

/**
 * A gauge is a metric that represents a single numerical value that can arbitrarily go up and down.
 */
export class Gauge<T extends string> {
  /**
   * @param configuration Configuration when creating a Gauge metric. Name and Help is mandatory
   */
  constructor(configuration: GaugeConfiguration<T>);

  /**
   * Increment gauge for given labels
   * @param labels Object with label keys and values
   * @param value The value to increment with
   */
  inc(labels: LabelValues<T>, value?: number): void;

  /**
   * Increment gauge
   * @param value The value to increment with
   */
  inc(value?: number): void;

  /**
   * Decrement gauge
   * @param labels Object with label keys and values
   * @param value Value to decrement with
   */
  dec(labels: LabelValues<T>, value?: number): void;

  /**
   * Decrement gauge
   * @param value The value to decrement with
   */
  dec(value?: number): void;

  /**
   * Set gauge value for labels
   * @param labels Object with label keys and values
   * @param value The value to set
   */
  set(labels: LabelValues<T>, value: number): void;

  /**
   * Set gauge value
   * @param value The value to set
   */
  set(value: number): void;

  /**
   * Set gauge value to current epoch time in ms
   * @param labels Object with label keys and values
   */
  setToCurrentTime(labels?: LabelValues<T>): void;

  /**
   * Start a timer where the gauges value will be the duration in seconds
   * @param labels Object with label keys and values
   * @return Function to invoke when timer should be stopped
   */
  startTimer(labels?: LabelValues<T>): (labels?: LabelValues<T>) => void;

  /**
   * Return the child for given labels
   * @param values Label values
   * @return Configured gauge with given labels
   */
  labels(...values: string[]): Gauge.Internal<T>;

  /**
   * Reset gauge values
   */
  reset(): void;

  /**
   * Remove metrics for the given label values
   * @param values Label values
   */
  remove(...values: string[]): void;
}

export namespace Gauge {
  interface Internal<T extends string> {
    /**
     * Increment gauge with value
     * @param value The value to increment with
     */
    inc(value?: number): void;

    /**
     * Decrement with value
     * @param value The value to decrement with
     */
    dec(value?: number): void;

    /**
     * Set gauges value
     * @param value The value to set
     */
    set(value: number): void;

    /**
     * Set gauge value to current epoch time in ms
     */
    setToCurrentTime(): void;

    /**
     * Start a timer where the gauges value will be the duration in seconds
     * @return Function to invoke when timer should be stopped
     */
    startTimer(): (labels?: LabelValues<T>) => void;
  }
}

export interface HistogramConfiguration<T extends string> {
  name: string;
  help: string;
  labelNames?: T[];
  buckets?: number[];
  registers?: Registry[];
  aggregator?: Aggregator;
}

/**
 * A histogram samples observations (usually things like request durations or response sizes) and counts them in configurable buckets
 */
export class Histogram<T extends string> {
  /**
   * @param configuration Configuration when creating the Histogram. Name and Help is mandatory
   */
  constructor(configuration: HistogramConfiguration<T>);

  /**
   * Observe value
   * @param value The value to observe
   */
  observe(value: number): void;
  /**
   * Observe value for given labels
   * @param labels Object with label keys and values
   * @param value The value to observe
   */
  observe(labels: LabelValues<T>, value: number): void;

  /**
   * Start a timer where the value in seconds will observed
   * @param labels Object with label keys and values
   * @return Function to invoke when timer should be stopped
   */
  startTimer(labels?: LabelValues<T>): (labels?: LabelValues<T>) => void;

  /**
   * Reset histogram values
   */
  reset(): void;

  /**
   * Return the child for given labels
   * @param values Label values
   * @return Configured histogram with given labels
   */
  labels(...values: string[]): Histogram.Internal<T>;

  /**
   * Remove metrics for the given label values
   * @param values Label values
   */
  remove(...values: string[]): void;
}

export namespace Histogram {
  interface Internal<T extends string> {
    /**
     * Observe value
     * @param value The value to observe
     */
    observe(value: number): void;

    /**
     * Start a timer where the value in seconds will observed
     * @param labels Object with label keys and values
     * @return Function to invoke when timer should be stopped
     */
    startTimer(): (labels?: LabelValues<T>) => void;
  }

  interface Config {
    /**
     * Buckets used in the histogram
     */
    buckets?: number[];
  }
}

export interface SummaryConfiguration<T extends string> {
  name: string;
  help: string;
  labelNames?: T[];
  percentiles?: number[];
  registers?: Registry[];
  aggregator?: Aggregator;
  maxAgeSeconds?: number;
  ageBuckets?: number;
  compressCount?: number;
}

/**
 * A summary samples observations
 */
export class Summary<T extends string> {
  /**
   * @param configuration Configuration when creating Summary metric. Name and Help is mandatory
   */
  constructor(configuration: SummaryConfiguration<T>);

  /**
   * Observe value in summary
   * @param value The value to observe
   */
  observe(value: number): void;
  /**
   * Observe value for given labels
   * @param labels Object with label keys and values
   * @param value Value to observe
   */
  observe(labels: LabelValues<T>, value: number): void;

  /**
   * Start a timer where the value in seconds will observed
   * @param labels Object with label keys and values
   * @return Function to invoke when timer should be stopped
   */
  startTimer(labels?: LabelValues<T>): (labels?: LabelValues<T>) => void;

  /**
   * Reset all values in the summary
   */
  reset(): void;

  /**
   * Return the child for given labels
   * @param values Label values
   * @return Configured summary with given labels
   */
  labels(...values: string[]): Summary.Internal<T>;

  /**
   * Remove metrics for the given label values
   * @param values Label values
   */
  remove(...values: string[]): void;
}

export namespace Summary {
  interface Internal<T extends string> {
    /**
     * Observe value in summary
     * @param value The value to observe
     */
    observe(value: number): void;

    /**
     * Start a timer where the value in seconds will observed
     * @param labels Object with label keys and values
     * @return Function to invoke when timer should be stopped
     */
    startTimer(): (labels?: LabelValues<T>) => void;
  }

  interface Config {
    /**
     * Configurable percentiles, values should never be greater than 1
     */
    percentiles?: number[];
  }
}

/**
 * Push metrics to a Pushgateway
 */
export class Pushgateway {
  /**
   * @param url Complete url to the Pushgateway. If port is needed append url with :port
   * @param options Options
   * @param registry Registry
   */
  constructor(url: string, options?: any, registry?: Registry);

  /**
   * Add metric and overwrite old ones
   * @param params Push parameters
   * @param callback Callback when request is complete
   */
  pushAdd(params: Pushgateway.Parameters, callback: (error?: Error, httpResponse?: any, body?: any) => void): void;

  /**
   * Overwrite all metric (using PUT to Pushgateway)
   * @param params Push parameters
   * @param callback Callback when request is complete
   */
  push(params: Pushgateway.Parameters, callback: (error?: Error, httpResponse?: any, body?: any) => void): void;

  /**
   * Delete all metrics for jobName
   * @param params Push parameters
   * @param callback Callback when request is complete
   */
  delete(params: Pushgateway.Parameters, callback: (error?: Error, httpResponse?: any, body?: any) => void): void;
}

export namespace Pushgateway {
  interface Parameters {
    /**
     * Jobname that is pushing the metric
     */
    jobName: string;
    /**
     * Label sets used in the url when making a request to the Pushgateway,
     */
    groupings?: {
      [key: string]: string;
    };
  }
}